requirements.txt
cffi == 1.16.0
defusedxml == 0.7.1
numpy == 1.26.4
opencv-contrib-python == 4.9.0.80
opencv-python  == 4.9.0.80
packaging == 23.2
pillow == 10.2.0
pip  == 24.0
pycparser == 2.21
PyGetWindow == 0.0.9
PyRect  == 0.2.0
setuptools  == 65.5.0